backend_fastapi/requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.2
openai>=1.3.7
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
typing-extensions>=4.7.0
supabase>=2.15.0
